Added physics documentation for ion_phys
I have written a document which is intended to define all our conventions and has the useful derivations of electric dipole, magentic dipole, and electric quadrupole transition matrix elements and rates. This is not a final document and is intended to be dynamic as we refine the code-base.

I have also added to the .gitignore file so that LaTex build files are not tracked.